indexing-service/requirements.txt
fastapi
uvicorn
qdrant-client==1.15.1
boto3
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests

pytest>=7.0.0
pytest-cov>=4.0.0 